requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-google-genai
google-generativeai
langsmith
langchain_community
langserve
fastapi
uvicorn
pydantic
langchain-ollama
langchain-openai
openai
langchain-anthropic
anthropic
langchain-huggingface
huggingface 
huggingface_hub
transformers
numpy
scikit-learn
faiss-cpu 
pypdf2
beautifulsoup4
requests
serpapi
PyPDF2
chromadb
openai
langchain
python-multipart
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
email-validator   # optional if you validate emails
apscheduler
sqlalchemy
websockets
